bots/rhizoo-alpha-bot/requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
pydantic>=2.0.0
loguru>=0.7.0
python-dotenv>=1.0.0
numpy>=1.24.0
